+++ templates/home.html
@@ -10,7 +10,7 @@
     <div class="container d-flex flex-column justify-content-end" style="min-height: 80vh; position: relative;">
         {% if user.is_staff %}
         <div style="position: absolute; top: 20px; right: 20px; z-index: 10;">
-            <a href="{% url 'travel:banner-upload' %}" class="btn btn-sm btn-outline-secondary">
+            <a href="{% url 'accounts:banner-upload' %}" class="btn btn-sm btn-outline-secondary">
                 Edit Banner
             </a>
         </div>
